+++ cmd/mq_json/main.go
@@ -109,7 +109,10 @@ func main() {
 		mqmetric.QueueManagerInitAttributes()
 		mqmetric.UsageInitAttributes()
 		mqmetric.ClusterInitAttributes()
+		mqmetric.ChannelAMQPInitAttributes()
+
 	}
+
 	// What metrics can the queue manager provide? Find out, and
 	// subscribe.
 	if err == nil {
@@ -127,6 +130,8 @@ func main() {
 		err = mqmetric.DiscoverAndSubscribe(discoverConfig)
 		// Also get the static attributes for any configured channels
 		mqmetric.RediscoverAttributes(ibmmq.MQOT_CHANNEL, config.cf.MonitoredChannels)
+		mqmetric.RediscoverAttributes(mqmetric.OT_CHANNEL_AMQP, config.cf.MonitoredAMQPChannels)
+
 	}
 
 	// Go into main loop for sending data to stdout